[PyTorch Edge] Add test for lite interpreter operator caching
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62306

Test to see if caching of operators works as expected. When caching operators during model load we look up using the operator name. This test ensures that even if there are multiple operators with the same name (in the same model), the caching distinguishes between the ones that have a different number of arguments specified during the call in the serialized bytecode.

In this specific test, there's a model with 3 methods, 2 of which return a `float32` tensor and one which return an `int64` dtype. Please see the comments in the diff for details.

#accept2ship

ghstack-source-id: 134529657

Differential Revision: [D29929116](https://our.internmc.facebook.com/intern/diff/D29929116/)